+++ crates/miden-protocol/src/account/builder/mod.rs
@@ -113,7 +113,7 @@ impl AccountBuilder {
     /// Adds a designated authentication [`AccountComponent`] to the builder.
     ///
     /// This component may contain multiple procedures, but is expected to contain exactly one
-    /// authentication procedure (named `auth_*`).
+    /// authentication procedure (marked with the `@auth_script` attribute).
     /// Calling this method multiple times will override the previous auth component.
     ///
     /// Procedures from this component will be placed at the beginning of the account procedure